+++ CHANGELOG.md
@@ -3,11 +3,15 @@
 * Add reduce and set functions
 * add while, ++ and -- macros
 * ignore comments everything after ; but not inside strings
+* gensym function
+* better string function
+
 
 ### Bugs
 * fix parsing empty strings
 * fix various errors catch by lint
 * fix parsing ALists with list as keys and values
+* fix parsing quasiquote that evaluate to single pair out if unquote
 
 ## 0.1.0
 * Initial version